Add files via upload
And here's an example of an automatic gate's control circuit's code, you can use it on your home gate or read it to undestando some coding tecnhniques as using timers instead of a delay and the "switch case machine" used to control the system. Try enjoy it. It still has some errors but is minimally debugged to work.